Added "boot settings" preferences
* Requires "Cyanogenmod" logo-themed u-boot.bin
* "Always Boot from eMMC" setting (/rom/u-boot.default)
* "Always Use Alternate Boot" setting (/rom/u-boot.altboot)
* Some random hints.

Seems to work, and believe it or not I think I actually did this the right way.  Logs extensively.

Note that the preferences are re-read from /rom when the program is first started.  This means that any stored preferences are ignored (in case files were changed manually).
The preference defaults are only updated on app's first-run (they could be made to reload when the "Boot Settings" screen activity is re-loaded or any time, but this is probably overkill).  This means that if you delete or change the files WHILE the program's running, it won't realize it until the app is re-run.

Even though /rom is mounted read/write, any app reading or writing there does need root privs.  The app tries w/o it, but should fallback to asking for root perms.

Please test extensively.